pyproject.toml
[project]
name = "log-scale-chronology"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "AcipenserSturio", email = "AcipenserSturio@users.noreply.github.com"},
]
dependencies = [
    "Pillow>=10.3.0",
]
requires-python = "==3.11.*"
readme = "README.md"
license = {text = "GPL-3.0"}

[tool.pdm]
distribution = true

[tool.pdm.scripts]
build = "python -m log_scale_chronology"
